api/requirements.txt
pyasyncore==1.0.4
python-dotenv==0.21.0
Flask==3.0.3
cassandra-driver==3.25.0